# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Tally API
  version: 0.1.0
  description: 家庭记账后端契约。金额统一为整数（分），日期为 YYYY-MM-DD，时间为 ISO-8601。
servers:
  - url: https://api.example.com
    description: 生产（HTTPS，实际域名由部署方替换；开发期可用本地 HTTPS 或局域网）
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      operationId: register
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "200":
          description: 注册成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: 新令牌
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
  /api/v1/auth/request-code:
    post:
      tags: [auth]
      operationId: requestCode
      summary: 发送短信验证码（开发模式回传 code，生产不返回）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string }
      responses:
        "200":
          description: 已发送
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
                  code: { type: string, nullable: true }
        "503": { $ref: "#/components/responses/Error" }
  /api/v1/auth/login-code:
    post:
      tags: [auth]
      operationId: loginWithCode
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string }
                code: { type: string }
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
  /api/v1/auth/forgot-password:
    post:
      tags: [auth]
      operationId: forgotPassword
      summary: 申请密码重置（生产模式不回传 resetToken，仅返回 ok）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string }
      responses:
        "200":
          description: 已受理。开发模式回传 resetToken，生产模式不返回。
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
                  resetToken: { type: string, nullable: true, description: 仅开发模式回传；生产环境恒为缺省 }
        "429": { $ref: "#/components/responses/Error" }
  /api/v1/auth/reset-password:
    post:
      tags: [auth]
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [resetToken, newPassword]
              properties:
                resetToken: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        "200":
          description: 重置成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "400": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: 统一错误
      headers:
        x-request-id:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    AuthResponse:
      type: object
      required: [user, token, refreshToken]
      properties:
        user: { $ref: "#/components/schemas/User" }
        token: { type: string }
        refreshToken: { type: string }
    User:
      type: object
      required: [id, email, displayName, createdAt]
      properties:
        id: { type: string }
        email: { type: string }
        displayName: { type: string }
        createdAt: { type: string }
    RegisterRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email: { type: string }
        password: { type: string }
        displayName: { type: string }
...